Configuring the auth endpoints on the resource server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 // using AutoMapper.QueryableExtensions;
@@ -51,12 +52,13 @@ public class AuctionsController : ControllerBase
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction=_mapper.Map<Auction>(auctionDto);
 
-            auction.Seller="test";
+            auction.Seller=User.Identity.Name;
             
             _context.Auctions.Add(auction);
             var newAuction=_mapper.Map<AuctionDto>(auction);
@@ -72,6 +74,7 @@ public class AuctionsController : ControllerBase
             );
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -81,6 +84,7 @@ public class AuctionsController : ControllerBase
             if(auction==null) return NotFound();
 
             //TODO: check seller==username
+            if(auction.Seller!= User.Identity.Name) return Forbid();
 
             auction.Item.Make=updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model=updateAuctionDto.Model ?? auction.Item.Model;
@@ -95,6 +99,7 @@ public class AuctionsController : ControllerBase
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)  
         {
@@ -103,6 +108,8 @@ public class AuctionsController : ControllerBase
             if(auction==null) return NotFound();
 
             //todo: check seller = username
+            if(auction.Seller!= User.Identity.Name) return Forbid();
+            
             _context.Auctions.Remove(auction);
             await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
